requirements.txt
Flask>=3.0.0
PyMySQL>=1.1.0
python-dotenv>=1.0.0
cryptography>=42.0.0
